Add files via upload
Updating Project 1: Modifying function isPuzzleSolvable() to add additional rule for determining whether a puzzle is solvable.
Initially I thought that when an initial puzzle had an odd number of inversions, it would determine the puzzle to be unsolvable. However, I realized while testing other puzzles (start and goal) that a puzzle being solvable depends on both initial and goal puzzles to be either both even or odd. This change is being made on this update.